% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xrf.R
\name{read_olympus_vanta}
\alias{read_olympus_vanta}
\alias{read_panalytical_txt}
\title{Read Olympus Vanta XRF files}
\usage{
read_olympus_vanta(path, extra_columns = "info")

read_panalytical_txt(path)
}
\arguments{
\item{path}{The location of the file}

\item{extra_columns}{extra columns from the original file to include in the output}
}
\value{
A data.frame
}
\description{
The standard Olympus Vanta file presents all elemental concentrations in ppm, and
all errors as 1 standard deviation.
}
\examples{
read_olympus_vanta(system.file("xrf_files/olympus_vanta_test.csv", package = "paleoxrf"))
read_panalytical_txt(system.file("xrf_files/panalytical_test.txt", package = "paleoxrf"))

}
